Moved AnonymousUsage, AnonymousReference
