+++ language/src/main/grammars/de/monticore/lang/SysMLBasis.mc4
@@ -122,34 +122,6 @@ component grammar SysMLBasis
     {noSpace(2);} "#" MCQualifiedName ;
 
   /**
-   * Anonyme Usage, deren "Typ" im Sinne der SysML (also Beispielsweise "attribute" oder "part") sich erst durch
-   * die Spezialisationen (Typisierungen) ergibt.
-   */
-  symbol scope AnonymousUsage implements SysMLUsage =
-      Modifier UserDefinedKeyword* SysMLIdentifier? Name? SysMLCardinality?
-      Specialization+ DefaultValue?
-      ("{"
-        SysMLElement*
-      "}" | ";");
-
-  symbolrule AnonymousUsage =
-      types:de.monticore.types.check.SymTypeExpression*
-      in:boolean
-      out:boolean ;
-
-  symbol scope AnonymousReference implements SysMLUsage =
-       Modifier UserDefinedKeyword* SysMLIdentifier? src:MCQualifiedName SysMLCardinality?
-       Specialization* DefaultValue?
-       ("{"
-          SysMLElement*
-        "}" | ";");
-
-  symbolrule AnonymousReference =
-      types:de.monticore.types.check.SymTypeExpression*
-      in:boolean
-      out:boolean ;
-
-  /**
    * Occurrence usages can be used in successions.
    * Example: "first send a to b if b then perform myAction;"
    */
