+++ language/src/main/grammars/de/monticore/lang/SysMLv2.mc4
@@ -171,6 +171,34 @@ grammar SysMLv2 extends SysMLExpressions,
      "}" | ";");
 
   /**
+     * Anonyme Usage, deren "Typ" im Sinne der SysML (also Beispielsweise "attribute" oder "part") sich erst durch
+     * die Spezialisationen (Typisierungen) ergibt.
+     */
+  symbol scope AnonymousUsage implements SysMLUsage =
+      Modifier UserDefinedKeyword* SysMLIdentifier? Name? SysMLCardinality?
+      Specialization+ DefaultValue?
+      ("{"
+        SysMLElement*
+      "}" | ";");
+
+  symbolrule AnonymousUsage =
+      types:de.monticore.types.check.SymTypeExpression*
+      in:boolean
+      out:boolean ;
+
+  symbol scope AnonymousReference implements SysMLUsage =
+       Modifier UserDefinedKeyword* SysMLIdentifier? src:MCQualifiedName SysMLCardinality?
+       Specialization* DefaultValue?
+       ("{"
+          SysMLElement*
+        "}" | ";");
+
+  symbolrule AnonymousReference =
+      types:de.monticore.types.check.SymTypeExpression*
+      in:boolean
+      out:boolean ;
+
+  /**
    * [An event occurence] is related to another occurrence usage,
    * representing the occurring event, by a reference subsetting relationship, which is a special kind of subsetting
    * relationship specified using the keyword references or the symbol ::>.
