--- language/src/main/java/de/monticore/lang/sysmlbasis/_ast/ASTAnonymousReference.java
@@ -1,10 +1,0 @@
-package de.monticore.lang.sysmlbasis._ast;
-
-public class ASTAnonymousReference extends ASTAnonymousReferenceTOP {
-
-  @Override
-  public String getName() {
-    return getSrc().getQName();
-  }
-
-}
