+++ language/src/main/java/de/monticore/lang/sysmlv2/_symboltable/AnonymousReferenceSymbolDeSer.java
@@ -1,5 +1,5 @@
 /* (c) https://github.com/MontiCore/monticore */
-package de.monticore.lang.sysmlbasis._symboltable;
+package de.monticore.lang.sysmlv2._symboltable;
 
 import de.monticore.lang.sysmlbasis.symboltable.SerializationUtil;
 import de.monticore.lang.sysmlv2.SysMLv2Mill;
@@ -9,10 +9,10 @@ import de.monticore.types.check.SymTypeExpressionDeSer;
 
 import java.util.List;
 
-public class AnonymousUsageSymbolDeSer extends AnonymousUsageSymbolDeSerTOP {
+public class AnonymousReferenceSymbolDeSer extends AnonymousReferenceSymbolDeSerTOP {
 
   @Override
-  protected void serializeTypes(List<SymTypeExpression> types, SysMLBasisSymbols2Json s2j) {
+  protected void serializeTypes(List<SymTypeExpression> types, SysMLv2Symbols2Json s2j) {
     SymTypeExpressionDeSer.serializeMember(s2j.getJsonPrinter(), "types", types);
   }
