+++ language/src/main/java/de/monticore/lang/sysmlv2/symboltable/completers/DirectionCompleter.java
@@ -1,7 +1,8 @@
 package de.monticore.lang.sysmlv2.symboltable.completers;
 
-import de.monticore.lang.sysmlbasis._ast.ASTAnonymousReference;
-import de.monticore.lang.sysmlbasis._ast.ASTAnonymousUsage;
+import de.monticore.lang.sysmlv2._ast.ASTAnonymousReference;
+import de.monticore.lang.sysmlv2._ast.ASTAnonymousUsage;
+import de.monticore.lang.sysmlv2._visitor.SysMLv2Visitor2;
 import de.monticore.lang.sysmlbasis._visitor.SysMLBasisVisitor2;
 import de.monticore.lang.sysmlparts._ast.ASTAttributeUsage;
 import de.monticore.lang.sysmlparts._visitor.SysMLPartsVisitor2;
@@ -10,7 +11,7 @@ import de.monticore.umlmodifier._ast.ASTModifier;
 /**
  * Sets the isIn / isOut properties of symbols based on Modifiers parsed to the AST.
  */
-public class DirectionCompleter implements SysMLBasisVisitor2, SysMLPartsVisitor2 {
+public class DirectionCompleter implements SysMLBasisVisitor2, SysMLPartsVisitor2, SysMLv2Visitor2{
 
   /**
    * Returns whether this is an input. Defaults to {@code true} if no direction was explicitly set. The keyword "inout"
